RS-11 and RS-12 order flow (#9)
Passenger can create order. Driver and accept/reject order. If accepted, at the end driver has to confirm the order. If order is rejected and timed out, it is retried 4 more times (total of 5 tries). If no driver is found, order is cancelled.

Also the requests are secured with JWT token